+++ apis/apigatewayv2/v1beta1/zz_api_types.go
@@ -227,7 +227,7 @@ type CorsConfigurationInitParameters struct {
 	ExposeHeaders []*string `json:"exposeHeaders,omitempty" tf:"expose_headers,omitempty"`
 
 	// Number of seconds that the browser should cache preflight request results.
-	MaxAge *int64 `json:"maxAge,omitempty" tf:"max_age,omitempty"`
+	MaxAge *float64 `json:"maxAge,omitempty" tf:"max_age,omitempty"`
 }
 
 type CorsConfigurationObservation struct {
@@ -248,7 +248,7 @@ type CorsConfigurationObservation struct {
 	ExposeHeaders []*string `json:"exposeHeaders,omitempty" tf:"expose_headers,omitempty"`
 
 	// Number of seconds that the browser should cache preflight request results.
-	MaxAge *int64 `json:"maxAge,omitempty" tf:"max_age,omitempty"`
+	MaxAge *float64 `json:"maxAge,omitempty" tf:"max_age,omitempty"`
 }
 
 type CorsConfigurationParameters struct {
@@ -275,7 +275,7 @@ type CorsConfigurationParameters struct {
 
 	// Number of seconds that the browser should cache preflight request results.
 	// +kubebuilder:validation:Optional
-	MaxAge *int64 `json:"maxAge,omitempty" tf:"max_age,omitempty"`
+	MaxAge *float64 `json:"maxAge,omitempty" tf:"max_age,omitempty"`
 }
 
 // APISpec defines the desired state of API
