+++ apis/docdb/v1beta1/zz_clustersnapshot_types.go
@@ -43,7 +43,7 @@ type ClusterSnapshotObservation struct {
 	KMSKeyID *string `json:"kmsKeyId,omitempty" tf:"kms_key_id,omitempty"`
 
 	// Port that the DocumentDB cluster was listening on at the time of the snapshot.
-	Port *int64 `json:"port,omitempty" tf:"port,omitempty"`
+	Port *float64 `json:"port,omitempty" tf:"port,omitempty"`
 
 	SnapshotType *string `json:"snapshotType,omitempty" tf:"snapshot_type,omitempty"`
